Add `#each_page` to ResourceCollection
each_page is similar to invoking each, except yields a page object
which is a collection of resources retrieved in a single API request
from the endpoint, along with a URL to the next page of results.

This allows consuming code to have more visibility into the pagination
process, for example to store the next page link in order to resume
pagination from there on error.